frontend.staging/requirements.txt
#requirements.txt
streamlit
psycopg2-binary
sqlalchemy
bcrypt
cryptography
infisical-python
todoist_api_python
